Switched to external CSM plugin
The camera-stage-mapping extension is now included from
an external repository, and removed from this one.
This should mean the code is more visible and useful
outside the server.  I've added
some improvements in the new
version of the plugin too.

NB we should make sure to freeze the version somehow before
releasing this version.