Refactored code out of app.py
There were a few things, notably logging and OpenAPI description
generation, that didn't really need to be in app.py.
I have moved these to the relevant modules, creating
a new module to set up logging.

The logging module sets things up at import time - I don't like
this, but it seems to be the pattern for WSGI applications,
and is the same as the current behaviour.

I've also moved adding the views to the app into a function
defined in views.__init__ - this doesn't change any
functionality.