Compliant entry point names
It's recommended to use entry point names starting with a PyPI
name and a dot, to avoid clashes.  I've settled on
openflexure_microscope for
the microscope-specific classes (stage/camera)
and labthings.extensions for extensions.
We probably want to support extensions that are not exclusively for
OFM - this feels like a reasonable namespace to use
either way, because any OFM extension must also
be a labthings extension.