Move atexit code into microscope class
Previously, the app module set up auto-closing of the hardware.
Now, this is done when we initialise the microscope.
The shutdown function is unregistered if the microscope
is closed manually.
NB this may mean the microscope won't get garbage collected
so we may have to be more careful about closing it if
we need to recreate it in the future.