Refactored app to allow reloading/fallback
I've split up the app module (which was something of
a behemoth) into several submodules.  This means there
is one module, "app", that actually instantiates the app.

Submodules handle the code that creates the app, a reloading
wrapper (that maintains WSGI compatibility but also allows
us to swap out the app object), and provision for
a fallback app if the main app can't load.

Note that reloading the main app once it has been
instantiated successfully is likely to go wrong, because there
are not yet sufficient safeguards to make sure everything is
properly shut down and released, before the replacement
app is created.

Extension loading is also split out into another module for
clarity and ease of error handling.