Eliminate import-time side effects
Setting things up at import-time is a nasty habit of
Flask apps.  It's necessary to do this in the main app
module so we can work with WSGI properly.
However, everywhere else it would be really helpful
to avoid side effects of importing modules.

There are now functions that initialise the paths and logging, which
must be called before starting the server.