Avoid defining default capture path at import
In an effort to minimise the amount of configuration that's done
at import time, I am no longer setting a global constant
with the default capture path.
This global variable is misleading, as the default
capture path is set in the CaptureManager
object and may not be the same as
DEFAULT_CAPTURE_PATH anyway.