Try using the OS image we most recently built
I've written a script to convert our SD card images into Docker
images. These are uploaded to our registry,
and the CI now runs tests in that image.
This should be as close as we can get to
the real Pi environment (though systemd is
not yet running, which probably isn't entirely a bad thing).