Don't cache the virtual environment
Caching the virtual environment can cause
problems if I need to overwrite it - it's simplest
just to re-generate the environment each run.
This will be done using cached downloads if
Pipfile.lock hasn't changed.

Caches are not guaranteed, so even if I did cache
the virtual environment, I would still need to
spin up a job to copy the cache into an
artifact to pass to later jobs.