Merge in IHI branch
The IHI branch includes several UI and scanning improvements
for blood smear imaging.
This branch merges in those changes to the developing v3 codebase.
The intention is that we will add switches to preserve
both the general-purpose features and the blood smear specific ones.
For now, it's blood smear only, until we get this version working.