Update camera protocol
The change should mean it's now possible to run a smart scan in a unit test. This is started in `test_dummy_server.py`.
Unfortunately, there are so many dependencies (specifically USB storage and
openflexure-stitching) that
it doesn't yet work with simulated hardware.
I anticipate we'd need defocus and possibly sphere-cap motion
before everything works as expected - the
smart scan is currently too smart for our test suite.