Use the raw/processing logic in the camera
smart_scan was previously doing all the image processing
to convert raw images into processed JPEGS.
This logic is now present in the camera Thing, and is used instead.

The change should mean it's now possible to run a smart scan in a unit test. This is started in `test_dummy_server.py`.
Unfortunately, there are so many dependencies (specifically USB storage and
openflexure-stitching) that
it doesn't yet work with simulated hardware.
I anticipate we'd need defocus and possibly sphere-cap motion
before everything works as expected - the
smart scan is currently too smart for our test suite.