Adding a native rgb2luv function to remove need for OpenCV on Pi
The native functions have been taken from scikit-image. scikit-image
(or OpenCV) are huge dependencies for the simple task of colourspace
conversion.

This is experimental, and needs testing for peformance and accuracy.

If it works it removes the need to install OpenCV on the Pi. OpenCV
will then only be used by the simulation camera and the OpenCV camera.